+++ src/main/java/com/supermartijn642/chunkloaders/ChunkLoaders.java
@@ -116,7 +116,6 @@ public class ChunkLoaders {
         handler.addGenerator(ChunkLoadersLanguageGenerator::new);
         handler.addGenerator(ChunkLoadersLootTableGenerator::new);
         handler.addGenerator(ChunkLoadersRecipeGenerator::new);
-        handler.addGenerator(ChunkLoadersRecipeGenerator::new);
         handler.addGenerator(ChunkLoadersTagGenerator::new);
     }
 }