+++ changelog.md
@@ -1,2 +1,2 @@
 ### Chunk Loaders 1.2.3
-- Initial release of Chunk Loaders for Minecraft 1.19.3
+- Initial release of Chunk Loaders for Minecraft 1.20